+++ .github/workflows/runtime_build_and_test.yml
@@ -547,7 +547,7 @@ jobs:
       - name: Download artifacts for base revision
         run: |
           git fetch origin main
-          GH_TOKEN=${{ github.token }} scripts/release/download-experimental-build-ghaction.js --commit=$(git rev-parse origin/main)
+          GH_TOKEN=${{ github.token }} scripts/release/download-experimental-build.js --commit=$(git rev-parse origin/main)
           mv ./build ./base-build
       # TODO: The `download-experimental-build` script copies the npm
       # packages into the `node_modules` directory. This is a historical
