+++ scripts/release/download-experimental-build.js
@@ -2,31 +2,58 @@
 
 'use strict';
 
-const {join} = require('path');
+const {join, relative} = require('path');
+const {handleError} = require('./utils');
+const yargs = require('yargs');
+const clear = require('clear');
+const theme = require('./theme');
 const {
-  addDefaultParamValue,
-  getPublicPackages,
-  handleError,
-} = require('./utils');
+  downloadBuildArtifacts,
+} = require('./shared-commands/download-build-artifacts');
 
-const downloadBuildArtifacts = require('./shared-commands/download-build-artifacts');
-const parseParams = require('./shared-commands/parse-params');
-const printSummary = require('./download-experimental-build-commands/print-summary');
+const argv = yargs.wrap(yargs.terminalWidth()).options({
+  releaseChannel: {
+    alias: 'r',
+    describe: 'Download the given release channel.',
+    requiresArg: true,
+    type: 'string',
+    choices: ['experimental', 'stable'],
+    default: 'experimental',
+  },
+  commit: {
+    alias: 'c',
+    describe: 'Commit hash to download.',
+    requiresArg: true,
+    demandOption: true,
+    type: 'string',
+  },
+}).argv;
 
-const run = async () => {
+function printSummary(commit) {
+  const commandPath = relative(
+    process.env.PWD,
+    join(__dirname, '../download-experimental-build.js')
+  );
+
+  clear();
+
+  const message = theme`
+    {caution An experimental build has been downloaded!}
+
+    You can download this build again by running:
+    {path   ${commandPath}} --commit={commit ${commit}}
+  `;
+
+  console.log(message.replace(/\n +/g, '\n').trim());
+}
+
+const main = async () => {
   try {
-    addDefaultParamValue('-r', '--releaseChannel', 'experimental');
-
-    const params = await parseParams();
-    params.cwd = join(__dirname, '..', '..');
-    params.packages = await getPublicPackages(true);
-
-    await downloadBuildArtifacts(params);
-
-    printSummary(params);
+    await downloadBuildArtifacts(argv.commit, argv.releaseChannel);
+    printSummary(argv.commit);
   } catch (error) {
     handleError(error);
   }
 };
 
-run();
+main();
