+++ scripts/release/shared-commands/download-build-artifacts.js
@@ -1,36 +1,98 @@
-#!/usr/bin/env node
-
 'use strict';
 
+const {join} = require('path');
+const theme = require('../theme');
 const {exec} = require('child-process-promise');
 const {existsSync} = require('fs');
-const {join} = require('path');
-const {getArtifactsList, logPromise} = require('../utils');
-const theme = require('../theme');
+const {logPromise} = require('../utils');
 
-const run = async ({build, cwd, releaseChannel}) => {
-  const artifacts = await getArtifactsList(build);
-  const buildArtifacts = artifacts.find(entry =>
-    entry.path.endsWith('build.tgz')
+if (process.env.GH_TOKEN == null) {
+  console.log(
+    theme`{error Expected GH_TOKEN to be provided as an env variable}`
+  );
+  process.exit(1);
+}
+
+const OWNER = 'facebook';
+const REPO = 'react';
+const WORKFLOW_ID = 'runtime_build_and_test.yml';
+const GITHUB_HEADERS = `
+  -H "Accept: application/vnd.github+json" \
+  -H "Authorization: Bearer ${process.env.GH_TOKEN}" \
+  -H "X-GitHub-Api-Version: 2022-11-28"`.trim();
+
+function getWorkflowId() {
+  if (
+    existsSync(join(__dirname, `../../../.github/workflows/${WORKFLOW_ID}`))
+  ) {
+    return WORKFLOW_ID;
+  } else {
+    throw new Error(
+      `Incorrect workflow ID: .github/workflows/${WORKFLOW_ID} does not exist. Please check the name of the workflow being downloaded from.`
+    );
+  }
+}
+
+async function getWorkflowRunId(commit) {
+  const res = await exec(
+    `curl -L ${GITHUB_HEADERS} https://api.github.com/repos/${OWNER}/${REPO}/actions/workflows/${getWorkflowId()}/runs?head_sha=${commit}&branch=main&exclude_pull_requests=true`
   );
 
-  if (!buildArtifacts) {
+  const json = JSON.parse(res.stdout);
+  let workflowRun;
+  if (json.total_count === 1) {
+    workflowRun = json.workflow_runs[0];
+  } else {
+    workflowRun = json.workflow_runs.find(
+      run => run.head_sha === commit && run.head_branch === 'main'
+    );
+  }
+
+  if (workflowRun == null || workflowRun.id == null) {
     console.log(
-      theme`{error The specified build (${build}) does not contain any build artifacts.}`
+      theme`{error The workflow run for the specified commit (${commit}) could not be found.}`
     );
     process.exit(1);
   }
 
-  // Download and extract artifact
-  const {CIRCLE_CI_API_TOKEN} = process.env;
-  let header = '';
-  // Add Circle CI API token to request header if available.
-  if (CIRCLE_CI_API_TOKEN != null) {
-    header = '-H "Circle-Token: ${CIRCLE_CI_API_TOKEN}" ';
+  return workflowRun.id;
+}
+
+async function getArtifact(workflowRunId, artifactName) {
+  const res = await exec(
+    `curl -L ${GITHUB_HEADERS} https://api.github.com/repos/${OWNER}/${REPO}/actions/runs/${workflowRunId}/artifacts?per_page=100&name=${artifactName}`
+  );
+
+  const json = JSON.parse(res.stdout);
+  let artifact;
+  if (json.total_count === 1) {
+    artifact = json.artifacts[0];
+  } else {
+    artifact = json.artifacts.find(
+      _artifact => _artifact.name === artifactName
+    );
   }
+
+  if (artifact == null) {
+    console.log(
+      theme`{error The specified workflow run (${workflowRunId}) does not contain any build artifacts.}`
+    );
+    process.exit(1);
+  }
+
+  return artifact;
+}
+
+async function downloadArtifactsFromGitHub(commit, releaseChannel) {
+  const workflowRunId = await getWorkflowRunId(commit);
+  const artifact = await getArtifact(workflowRunId, 'artifacts_combined');
+
+  // Download and extract artifact
+  const cwd = join(__dirname, '..', '..', '..');
   await exec(`rm -rf ./build`, {cwd});
   await exec(
-    `curl -L $(fwdproxy-config curl) ${buildArtifacts.url} ${header}| tar -xvz`,
+    `curl -L ${GITHUB_HEADERS} ${artifact.archive_download_url} \
+    > a.zip && unzip a.zip -d . && rm a.zip build2.tgz && tar -xvzf build.tgz && rm build.tgz`,
     {
       cwd,
     }
@@ -59,17 +121,16 @@ const run = async ({build, cwd, releaseChannel}) => {
     process.exit(releaseChannel);
   }
   await exec(`cp -r ./build/${sourceDir} ./build/node_modules`, {cwd});
-};
+}
 
-module.exports = async ({build, commit, cwd, releaseChannel}) => {
-  let buildLabel;
-  if (commit !== null) {
-    buildLabel = theme`commit {commit ${commit}} (build {build ${build}})`;
-  } else {
-    buildLabel = theme`build {build ${build}}`;
-  }
+async function downloadBuildArtifacts(commit, releaseChannel) {
+  const label = theme`commit {commit ${commit}})`;
   return logPromise(
-    run({build, cwd, releaseChannel}),
-    theme`Downloading artifacts from Circle CI for ${buildLabel}`
+    downloadArtifactsFromGitHub(commit, releaseChannel),
+    theme`Downloading artifacts from GitHub for ${label}`
   );
+}
+
+module.exports = {
+  downloadBuildArtifacts,
 };
